fix(slack-edit): tolerant metadata parse + drop historical PR references
Follow-up to PR #26618.

- updateMessageContentAndMetadata: swap raw JSON.parse of messages.metadata
  for the tolerant safeParseRecord helper. Malformed legacy rows no longer
  throw after recordInbound has persisted the dedup row, so edits can't be
  permanently dropped on retry.
- edit-intercept.ts: rewrite three comments that referenced 'PR 11' /
  'pre-PR-11' / 'post-PR-11' to describe the current state of the code,
  per assistant/AGENTS.md's no-history-in-comments rule.
- Promote safeParseRecord to util/json.ts and remove the local copy in
  edit-intercept.ts so both call sites share the helper.
