+++ assistant/src/util/json.ts
@@ -8,3 +8,20 @@ export function parseJsonSafe<T = unknown>(text: string): T | null {
     return null;
   }
 }
+
+/** Tolerant JSON parse that returns `{}` for invalid or non-object payloads. */
+export function safeParseRecord(raw: string): Record<string, unknown> {
+  try {
+    const parsed = JSON.parse(raw);
+    if (
+      parsed === null ||
+      typeof parsed !== "object" ||
+      Array.isArray(parsed)
+    ) {
+      return {};
+    }
+    return parsed as Record<string, unknown>;
+  } catch {
+    return {};
+  }
+}
